service/requirements.txt
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
djangorestframework-jsonapi==2.1.1
inflection==0.3.1
psycopg2==2.7
